Move function into service
